client_generated.go
